requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
tortoise-orm>=0.20.0
aerich>=0.7.2
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.9
email-validator>=2.1.0
asyncpg>=0.29.0  # PostgreSQL driver
redis>=5.0.1     # For caching
pytest>=8.0.0    # For testing
pytest-asyncio>=0.23.0  # For async tests
httpx>=0.26.0    # For testing HTTP clients 